[CLEAN] packer complete code before fixes
